Update: use separate threads to submit jobs
This is a workaround for an issue where calls to cudaStreamAddCallback()
blocks. With multiple threads, the main thread can continue to create
and submit jobs to other GPUs even in the event that AddCallback()
blocks for a certain GPU.

It seems that there is an internal limit of 64 pending callbacks (per
device). The observed behaviour is that cudaStreamAddCallback() is very
fast up to the 64th call to it for a certain device. At the 65th call,
it will block, presumably until at least one job on the selected device
has finished.